python/demo/pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.56.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
